feat:calculator 분리
